feat(book) : book entity create 2
